config.def.h: Use st instead of kitty for scratchpad
Kitty for all it's advantages has one drawback, and that being it
cannot spawn in a floating small window. When kitty was used as the
scratchpad, it would spawn a window the size of a normal tiling window
taking up the whole screen, and I would have to manually resize it and
the change would not be persistent, and there is no way to configure
kitty to launch the way I want it to.

St however launches in precisely the way I want it to no configuration
required.